Refactor task tracker to workflow_utils package
Move task tracker functionality from utils to new workflow_utils package for better architectural separation. Update imports and tests to match new structure.
