+++ src/mcp_coder/workflow_utils/__init__.py
@@ -1,0 +1,25 @@
+"""Workflow utilities for automated development processes.
+
+This package provides tools for managing development workflows, including
+task tracking, project status monitoring, and automation helpers.
+"""
+
+from .task_tracker import (
+    TaskInfo,
+    TaskTrackerError,
+    TaskTrackerFileNotFoundError,
+    TaskTrackerSectionNotFoundError,
+    get_incomplete_tasks,
+    is_task_done,
+)
+
+__all__ = [
+    # Task tracker operations
+    "get_incomplete_tasks",
+    "is_task_done",
+    "TaskInfo",
+    # Exception types
+    "TaskTrackerError",
+    "TaskTrackerFileNotFoundError",
+    "TaskTrackerSectionNotFoundError",
+]
